store-api-sequelize: repository sale

+++ Módulo 3 - Bancos de Dados + Node.js/Bancos de Dados SQL com ORM/store-api-sequelize/services/product.service.js
@@ -19,6 +19,12 @@ async function getProduct(id) {
 }
 
 async function deleteProduct(id) {
+  const sales = await SaleRepository.getSalesByProductId(id);
+
+  if (sales.length > 0) {
+    throw new Error('Não é possível excluir o produto pois ele tem vendas.');
+  }
+
   await ProductRepository.deleteProduct(id);
 }
